# JSON encoding further optimization