--- a/json.go
+++ b/json.go
@@ -21,48 +21,10 @@ func appendJSONString(dst []byte, s string) []byte {
 		// Check if the character needs encoding. Control characters, slashes,
 		// and the double quote need json encoding. Bytes above the ascii
 		// boundary needs utf8 encoding.
-		if s[i] < ' ' || s[i] == '\\' || s[i] == '"' || s[i] > 126 {
-			// We encountered a character that needs to be encoded. Let's
-			// append the previous simple characters to the byte slice
-			// and switch our operation to read and encode the remainder
-			// characters byte-by-byte.
-			dst = append(dst, s[:i]...)
-			for i < len(s) {
-				if b := s[i]; b < utf8.RuneSelf {
-					switch b {
-					case '"', '\\':
-						dst = append(dst, '\\', b)
-					case '\b':
-						dst = append(dst, '\\', 'b')
-					case '\f':
-						dst = append(dst, '\\', 'f')
-					case '\n':
-						dst = append(dst, '\\', 'n')
-					case '\r':
-						dst = append(dst, '\\', 'r')
-					case '\t':
-						dst = append(dst, '\\', 't')
-					default:
-						if b >= 0x20 {
-							dst = append(dst, b)
-						} else {
-							dst = append(dst, '\\', 'u', '0', '0',
-								hex[b>>4], hex[b&0xF])
-						}
-					}
-					i++
-					continue
-				}
-				r, size := utf8.DecodeRuneInString(s[i:])
-				if r == utf8.RuneError && size == 1 {
-					dst = append(dst, `\ufffd`...)
-					i++
-					continue
-				}
-				dst = append(dst, s[i:i+size]...)
-				i += size
-			}
-			// End with a double quote
+		if s[i] < 0x20 || s[i] > 0x7e || s[i] == '\\' || s[i] == '"' {
+			// We encountered a character that needs to be encoded. Switch
+			// to complex version of the algorithm.
+			dst = appendJSONStringComplex(dst, s, i)
 			return append(dst, '"')
 		}
 	}
@@ -71,4 +33,64 @@ func appendJSONString(dst []byte, s string) []byte {
 	dst = append(dst, s...)
 	// End with a double quote
 	return append(dst, '"')
+}
+
+// appendJSONStringComplex is used by appendJSONString to take over an in
+// progress JSON string encoding that encountered a character that needs
+// to be encoded.
+func appendJSONStringComplex(dst []byte, s string, i int) []byte {
+	start := 0
+	for i < len(s) {
+		b := s[i]
+		if b >= utf8.RuneSelf {
+			r, size := utf8.DecodeRuneInString(s[i:])
+			if r == utf8.RuneError && size == 1 {
+				// In case of error, first append previous simple characters to
+				// the byte slice if any and append a remplacement character code
+				// in place of the invalid sequence.
+				if start < i {
+					dst = append(dst, s[start:i]...)
+				}
+				dst = append(dst, `\ufffd`...)
+				i += size
+				start = i
+				continue
+			}
+			i += size
+			continue
+		}
+		if b >= 0x20 && b <= 0x7e && b != '\\' && b != '"' {
+			i++
+			continue
+		}
+		// We encountered a character that needs to be encoded.
+		// Let's append the previous simple characters to the byte slice
+		// and switch our operation to read and encode the remainder
+		// characters byte-by-byte.
+		if start < i {
+			dst = append(dst, s[start:i]...)
+		}
+		switch b {
+		case '"', '\\':
+			dst = append(dst, '\\', b)
+		case '\b':
+			dst = append(dst, '\\', 'b')
+		case '\f':
+			dst = append(dst, '\\', 'f')
+		case '\n':
+			dst = append(dst, '\\', 'n')
+		case '\r':
+			dst = append(dst, '\\', 'r')
+		case '\t':
+			dst = append(dst, '\\', 't')
+		default:
+			dst = append(dst, '\\', 'u', '0', '0', hex[b>>4], hex[b&0xF])
+		}
+		i++
+		start = i
+	}
+	if start < len(s) {
+		dst = append(dst, s[start:]...)
+	}
+	return dst
 }