Optimize JSON encoding even further

Last optimization was for JSON string with no character to encode. This
version focuses on strings with some chars to encode, trying to apply
the same trick for substrings that do not need encoding.

```
benchmark                old ns/op     new ns/op    delta
.../EncodingLast-8       62.8          61.1         -2.71%
.../RuneFirst-8          164           129          -21.34%
.../RuneMiddle-8         133           96.9         -27.14%
.../RuneLast-8           81.9          73.5         -10.26%
.../NoEncoding-8         60.2          51.3         -14.78%
.../EncodingFirst-8      140           116          -17.14%
.../EncodingMiddle-8     112           86.4         -22.86%
```

cc @tidwall